@@ -1,35 +1,71 @@
 {
-  "name": "avweather-cli",
-  "description": "Terminal application for getting aviation/pilot weather",
-  "version": "0.4.3",
   "author": "David Fekke @davidfekke",
   "bin": {
     "avwx": "./bin/run"
   },
-  "bugs": "https://github.com/davidfekke/avweather-cli/issues",
   "dependencies": {
-    "@oclif/command": "^1.5.1",
-    "@oclif/config": "^1.7.6",
-    "@oclif/plugin-help": "^2.1.2",
-    "axios": "^0.18.1",
-    "fast-xml-parser": "^3.12.5"
+    "@oclif/core": "^3",
+    "@oclif/plugin-help": "^5",
+    "@oclif/plugin-plugins": "^4"
   },
+  "description": "Terminal application for getting aviation/pilot weather",
   "devDependencies": {
-    "@oclif/dev-cli": "^1.18.1",
-    "eslint": "^5.5.0",
-    "eslint-config-oclif": "^3.1.0",
-    "globby": "^8.0.1"
+    "@oclif/prettier-config": "^0.2.1",
+    "@oclif/test": "^3",
+    "@types/chai": "^4",
+    "@types/mocha": "^10",
+    "@types/node": "^18",
+    "chai": "^4",
+    "eslint": "^8",
+    "eslint-config-oclif": "^5",
+    "eslint-config-oclif-typescript": "^3",
+    "eslint-config-prettier": "^9.0.0",
+    "mocha": "^10",
+    "oclif": "^4.0.3",
+    "shx": "^0.3.4",
+    "ts-node": "^10.9.1",
+    "typescript": "^5"
   },
   "engines": {
-    "node": ">=14.0.0"
+    "node": ">=18.0.0"
   },
   "files": [
     "/bin",
-    "/npm-shrinkwrap.json",
-    "/oclif.manifest.json",
-    "/src"
+    "/dist",
+    "/oclif.manifest.json"
   ],
-  "homepage": "https://github.com/davidfekke/avweather-cli",
+  "homepage": "https://github.com/davidfekke/avweather-cli/issues",
+  "license": "Apache-2.0",
+  "main": "src/index.js",
+  "name": "avweather-cli",
+  "oclif": {
+    "bin": "avwx",
+    "dirname": "avweather-cli",
+    "commands": "./dist/commands",
+    "plugins": [
+      "@oclif/plugin-help",
+      "@oclif/plugin-plugins"
+    ],
+    "topicSeparator": " ",
+    "topics": {
+      "hello": {
+        "description": "Say hello to the world and others"
+      }
+    }
+  },
+  "repository": "davidfekke/avweather-cli",
+  "scripts": {
+    "build": "shx rm -rf dist && tsc -b",
+    "lint": "eslint . --ext .ts",
+    "postpack": "shx rm -f oclif.manifest.json",
+    "posttest": "npm run lint",
+    "prepack": "npm run build && oclif manifest && oclif readme",
+    "prepare": "npm run build",
+    "test": "mocha --forbid-only \"test/**/*.test.ts\"",
+    "version": "oclif readme && git add README.md"
+  },
+  "version": "0.5.0",
+  "bugs": "https://github.com/davidfekke/avweather-cli/issues",
   "keywords": [
     "oclif",
     "metar",
@@ -37,21 +73,7 @@
     "aviation",
     "weather"
   ],
-  "license": "Apache-2.0",
-  "main": "src/index.js",
-  "oclif": {
-    "commands": "./src/commands",
-    "bin": "avwx",
-    "plugins": [
-      "@oclif/plugin-help"
-    ]
-  },
-  "repository": "davidfekke/avweather-cli",
-  "scripts": {
-    "postpack": "rm -f oclif.manifest.json npm-shrinkwrap.json",
-    "posttest": "eslint .",
-    "prepack": "oclif-dev manifest && oclif-dev readme && npm shrinkwrap",
-    "test": "echo NO TESTS",
-    "version": "oclif-dev readme && git add README.md"
-  }
+  "types": "dist/index.d.ts",
+  "exports": "./lib/index.js",
+  "type": "module"
 }
